cambios en la impresion

+++ WPFMultired/Classes/Printer/DataPrinter.cs
@@ -15,9 +15,9 @@ namespace WPFMultired.Classes.Printer
 
         public Image image { get; set; }
 
-        public int x { get; set; }
+        public float x { get; set; }
 
-        public int y { get; set; }
+        public float y { get; set; }
 
         public Font font { get; set; }
